+++ Controllers/E131Controller/RGBController_E131.cpp
@@ -71,8 +71,8 @@ RGBController_E131::RGBController_E131(std::vector<E131Device> device_list)
 
     for(unsigned int device_idx = 0; device_idx < devices.size(); device_idx++)
     {
-        float universe_size = devices[device_idx].universe_size;
-        unsigned int total_universes = ceil( ( ( devices[device_idx].num_leds * 3 ) + devices[device_idx].start_channel ) / universe_size );
+        float universe_size = (float)devices[device_idx].universe_size;
+        unsigned int total_universes = (unsigned int)ceil( ( ( devices[device_idx].num_leds * 3 ) + devices[device_idx].start_channel ) / universe_size );
 
         for(unsigned int univ_idx = 0; univ_idx < total_universes; univ_idx++)
         {
@@ -157,8 +157,8 @@ RGBController_E131::RGBController_E131(std::vector<E131Device> device_list)
         /*-----------------------------------------*\
         | Add Universes                             |
         \*-----------------------------------------*/
-        float universe_size = devices[device_idx].universe_size;
-        unsigned int total_universes = ceil( ( ( devices[device_idx].num_leds * 3 ) + devices[device_idx].start_channel ) / universe_size );
+        float universe_size = (float)devices[device_idx].universe_size;
+        unsigned int total_universes = (unsigned int)ceil( ( ( devices[device_idx].num_leds * 3 ) + devices[device_idx].start_channel ) / universe_size );
 
         for (unsigned int univ_idx = 0; univ_idx < total_universes; univ_idx++)
         {
@@ -178,7 +178,7 @@ RGBController_E131::RGBController_E131(std::vector<E131Device> device_list)
                 e131_packet_t   packet;
                 e131_addr_t     dest_addr;
 
-                e131_pkt_init(&packet, universe, universe_size);
+                e131_pkt_init(&packet, (uint16_t)universe, (uint16_t)universe_size);
 
                 if(multicast)
                 {
@@ -384,8 +384,8 @@ void RGBController_E131::DeviceUpdateLEDs()
 
     for(std::size_t device_idx = 0; device_idx < devices.size(); device_idx++)
     {
-        float universe_size = devices[device_idx].universe_size;
-        unsigned int total_universes = ceil( ( ( devices[device_idx].num_leds * 3 ) + devices[device_idx].start_channel ) / universe_size );
+        float universe_size = (float)devices[device_idx].universe_size;
+        unsigned int total_universes = (unsigned int)ceil( ( ( devices[device_idx].num_leds * 3 ) + devices[device_idx].start_channel ) / universe_size );
         unsigned int channel_idx = devices[device_idx].start_channel;
         unsigned int led_idx = 0;
         unsigned int rgb_idx = 0;
